Implement Data page and React routing
